+++ src/features/groups/components/ContainerCardList/ContainerCard/ContainerCardDropdownMenu/index.tsx
@@ -5,7 +5,7 @@ import { IContainer } from '@/types/definition';
 
 import { useState } from 'react';
 
-import { ContainerCardDropdownMenuContent } from './ContainerCardDropdownMenuTriggerButton';
+import { ContainerCardDropdownMenuContent } from './ContainerCardDropdownMenuContent';
 import { DeleteContainerDialogContent } from './DeleteContainerDialogContent';
 
 interface IContainerCardDropdownMenuProps {
